+++ Editor/ToolsMenu.cs
@@ -1,19 +1,21 @@
+#if UNITY_EDITOR
 using System.IO;
+using ringo;
 using UnityEngine;
 using static UnityEditor.AssetDatabase;
-# if UNITY_EDITOR
 using UnityEditor;
-# endif
 
 namespace akira
 {
     public static class ToolsMenu
     {
-        [MenuItem("Tools/Setup/Create Default Folders")]
-        public static void CreateDefaultFolders()
+        private const string RootFolder = "_Project";
+        
+        [MenuItem("Tools/Setup/Folders/Type-Based")]
+        public static void CreateTypeBasedDefaultFolders()
         {
             Folders.CreateDirectories(
-                "_Project",
+                RootFolder,
                 "_Scripts>Controllers",
                 "_Scripts>Editor",
                 "_Scripts>Interfaces",
@@ -44,6 +46,56 @@ namespace akira
                 "Scenes>Temporary Scenes",
                 "Sprites>UI"
             );
+            Refresh();
+        }
+        
+        [MenuItem("Tools/Setup/Folders/Function-Based")]
+        public static void CreateFunctionBasedDefaultFolders()
+        {
+            Folders.CreateDirectories(
+                RootFolder,
+
+                "_Dev>FirstnameLastname",
+                "_Dev>_Lost&Found",
+
+                "_Scripts>Controllers",
+                "_Scripts>Editor",
+                "_Scripts>Interfaces",
+                "_Scripts>Managers",
+                "_Scripts>Objects",
+                "_Scripts>Scriptables",
+                "_Scripts>Spawners",
+                "_Scripts>States",
+                "_Scripts>Systems",
+                "_Scripts>UI",
+                "_Scripts>Units",
+                "_Scripts>Utilities",
+
+                "Gameplay>Triggers", 
+                "Gameplay>Interactibles", 
+                "Gameplay>Pickups", 
+                "Gameplay>Obstacles", 
+
+                "Audio>Music",
+                "Audio>SFX",
+
+                "Levels", 
+
+                "UI>Fonts",
+
+                "Resources"
+            );
+            
+            WorldObjectFolders.Create(
+                RootFolder, 
+                "Objects>Architecture>[ArchitectureName]", 
+                "Objects>Props>[PropName]", 
+                "Characters>[CharacterName]", 
+                "Enemies>[EnemyName]", 
+                "VFX>[VFXName]",
+                "Player"
+                );
+            
             Refresh();
         }
 
@@ -79,3 +131,4 @@ namespace akira
         }
     }
 }
+#endif